refactor: extract FilterCoordinator from MainContentCoordinator
